# openapi.yaml
openapi: 3.0.3
info:
  title: FitLife API
  description: |
    REST API untuk aplikasi FitLife — platform nutrisi dan kesehatan berbasis Next.js.

    ## Autentikasi
    API ini mendukung dua metode autentikasi:
    1. **JWT Bearer Token** (direkomendasikan untuk Flutter/mobile) — kirim token di header `Authorization: Bearer <token>`
    2. **Cookie httpOnly** (untuk web browser) — cookie `token` di-set otomatis setelah login

    Token JWT berlaku selama **7 hari** dan berisi payload: `userId`, `role`, `email`.

    ## Base URL
    - **Lokal:** `http://localhost:3000`
    - **Produksi:** `https://<domain-anda>.vercel.app`
  version: "1.0.0"
  contact:
    name: FitLife Dev Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://your-domain.vercel.app
    description: Production server (Vercel)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Token JWT diperoleh dari response login/register. Gunakan di header:
        `Authorization: Bearer <token>`
  schemas:
...
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          example: "password_lama"
        newPassword:
          type: string
          minLength: 8
          example: "password_baru_123"

    # ── Menu ──────────────────────────────────────────────
    Menu:
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          example: "nasi-merah-dengan-sayur-1710000000000"
        nama_menu:
          type: string
          example: "Nasi Merah dengan Sayur"
        deskripsi:
          type: string
          example: "Menu sehat untuk diet"
        target_status:
          type: string
          enum: [Kurus, Normal, Berlebih, Obesitas]
          example: "Normal"
        kalori:
          type: number
          example: 350
        gambar_url:
          type: string
          nullable: true
          example: "https://res.cloudinary.com/..."
        dibaca:
          type: integer
          description: Jumlah kali dibaca (view count)
          example: 42
        created_at:
          type: string
          format: date-time
    MenuRequest:
      type: object
      required: [nama_menu, deskripsi, target_status, kalori]
      properties:
        nama_menu:
          type: string
          example: "Nasi Merah dengan Sayur"
        deskripsi:
          type: string
          example: "Menu sehat untuk diet"
        target_status:
          type: string
          enum: [Kurus, Normal, Berlebih, Obesitas]
          example: "Normal"
        kalori:
          type: number
          example: 350
        gambar_url:
          type: string
          nullable: true
          example: "https://res.cloudinary.com/..."
    PaginatedMenus:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Menu'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ── Artikel ───────────────────────────────────────────
...
    PerhitunganRequest:
      type: object
      required: [tinggi_badan, berat_badan]
      properties:
        tinggi_badan:
          type: number
          description: Tinggi badan dalam cm
          example: 170
        berat_badan:
          type: number
          description: Berat badan dalam kg
          example: 70
    PerhitunganResponse:
      type: object
      properties:
        bmi:
          type: number
          description: Nilai BMI (1 desimal)
          example: 24.2
        status:
          type: string
          enum: [Kurus, Normal, Berlebih, Obesitas]
          example: "Normal"
    PerhitunganHistory:
      type: object
      properties:
        id:
          type: integer
          example: 1
        user_id:
          type: integer
          example: 1
        tinggi_badan:
          type: number
          example: 170
        berat_badan:
          type: number
          example: 70
        bmi:
          type: number
          example: 24.2
        status:
          type: string
          example: "Normal"
        created_at:
          type: string
          format: date-time

    # ── Account (Admin) ───────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 10
        hasNextPage:
          type: boolean
          example: true
        hasPrevPage:
          type: boolean
          example: false
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Unauthorized"
    ValidationError:
      type: object
      properties:
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            email: ["Invalid email format"]
            password: ["Password minimal 8 karakter"]
paths:
...
  /api/profile/password:
    patch:
      tags: [Profile]
      summary: Ganti password user
      description: |
        Hanya tersedia untuk akun yang mendaftar dengan email/password
        (bukan Google OAuth).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password berhasil diubah
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password berhasil diubah"
        '400':
          description: Akun tidak menggunakan password
        '401':
          description: Unauthorized atau password saat ini salah
        '500':
          description: Internal Server Error

  # ============================================================
  # MENUS
  # ============================================================
  /api/menus:
    get:
      tags: [Menus]
      summary: Ambil daftar menu
      description: |
        Ambil semua menu makanan. Mendukung filter berdasarkan `target` dan `id`,
        serta **pagination opsional** dengan `page` dan `limit`.

        - Tanpa `page`/`limit`: mengembalikan array biasa (backward compatible)
        - Dengan `page`/`limit`: mengembalikan objek dengan `data` dan `meta`
      parameters:
        - name: target
          in: query
          description: Filter berdasarkan target status
          schema:
            type: string
            enum: [Kurus, Normal, Berlebih, Obesitas]
          example: Normal
        - name: id
          in: query
          description: Filter berdasarkan ID menu
          schema:
            type: integer
          example: 1
        - name: page
          in: query
          description: Nomor halaman (aktifkan mode pagination)
          schema:
            type: integer
            default: 1
          example: 1
        - name: limit
          in: query
          description: Jumlah item per halaman
          schema:
            type: integer
            default: 10
          example: 10
      responses:
        '200':
          description: Daftar menu
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Menu'
                    description: Response tanpa pagination
                  - $ref: '#/components/schemas/PaginatedMenus'
                    description: Response dengan pagination
        '400':
          description: Target status tidak valid atau ID tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error

    post:
      tags: [Menus]
      summary: Tambah menu baru
      description: |
        Menambahkan menu baru ke database. Slug di-generate otomatis dari `nama_menu`.

        > ⚠️ **Catatan Keamanan:** Endpoint ini belum dilindungi role admin.
        > Semua user (termasuk yang tidak login) bisa mengaksesnya.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuRequest'
      responses:
        '201':
          description: Menu berhasil ditambahkan
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Menu berhasil ditambahkan"
                  data:
                    $ref: '#/components/schemas/Menu'
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '500':
          description: Internal Server Error
  /api/menus/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        description: Slug unik menu
        schema:
          type: string
        example: "nasi-merah-dengan-sayur-1710000000000"

    get:
      tags: [Menus]
      summary: Detail menu berdasarkan slug
      description: Mengambil detail satu menu. View count (`dibaca`) akan otomatis bertambah.
      responses:
        '200':
          description: Data menu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          description: Menu tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error

    put:
      tags: [Menus]
      summary: Update menu (full update)
      description: |
        Update data menu secara penuh. Slug akan di-regenerate dari `nama_menu` baru.

        > ⚠️ **Catatan Keamanan:** Belum dilindungi role admin.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MenuRequest'
      responses:
        '200':
          description: Menu berhasil diperbarui
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Berhasil diperbarui"
                  data:
                    $ref: '#/components/schemas/Menu'
        '400':
          description: Validasi gagal
        '404':
          description: Menu tidak ditemukan
        '500':
          description: Internal Server Error

    delete:
      tags: [Menus]
      summary: Hapus menu
      description: |
        Menghapus menu berdasarkan slug.

        > ⚠️ **Catatan Keamanan:** Belum dilindungi role admin.
      responses:
        '200':
          description: Menu berhasil dihapus
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Menu berhasil dihapus"
        '404':
          description: Menu tidak ditemukan
        '500':
          description: Internal Server Error

  # ============================================================
  # ARTIKELS
  # ============================================================
...
  /api/perhitungan:
    post:
      tags: [Perhitungan]
      summary: Hitung BMI
      description: |
        Menghitung BMI berdasarkan tinggi dan berat badan.

        Jika user sedang login (token dikirim), hasil perhitungan akan otomatis
        disimpan ke riwayat. Jika tidak login, tetap mengembalikan hasil tanpa menyimpan.
      security:
        - BearerAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PerhitunganRequest'
      responses:
        '200':
          description: Hasil perhitungan BMI
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PerhitunganResponse'
              example:
                bmi: 24.2
                status: "Normal"
        '400':
          description: Data tidak lengkap atau tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error

    get:
      tags: [Perhitungan]
      summary: Riwayat perhitungan BMI user
      description: |
        Mengembalikan 10 riwayat perhitungan BMI terakhir untuk user yang sedang login.

        **Membutuhkan autentikasi.**
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Riwayat perhitungan BMI
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PerhitunganHistory'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal Server Error

  # ============================================================
  # ACCOUNTS (Admin)
  # ============================================================
...